all kinds of stuff

+++ src/app/reviews/containers/reviews-list/reviews-list.component.ts
@@ -1,5 +1,7 @@
 import { Component, Input } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Store } from '@ngrx/store';
+import { DialogConfirmComponent } from 'src/app/core/components/dialog-confirm/dialog-confirm.component';
 import { IReview } from 'src/app/shared/interfaces/review.interface';
 import { ReviewActions } from '../../state';
 
@@ -11,9 +13,19 @@ import { ReviewActions } from '../../state';
 export class ReviewsListComponent {
   @Input() reviews!: IReview[] | [] | null;
 
-  constructor(private store: Store) {}
+  constructor(private store: Store, public dialog: MatDialog) {}
 
   deleteReviewHandler(id: string) {
-    this.store.dispatch(ReviewActions.deleteReview({ reviewId: id }));
+    this.openDialog(id);
+  }
+
+  openDialog(id: string) {
+    const dialogRef = this.dialog.open(DialogConfirmComponent);
+
+    dialogRef.afterClosed().subscribe((confirmed) => {
+      if (confirmed) {
+        this.store.dispatch(ReviewActions.deleteReview({ reviewId: id }));
+      }
+    });
   }
 }
